Create File I\O Characters
This was an assignment I completed for my CS010B class, the instructions are attached below:

"Implement the function charCnt. charCnt is passed in a name of a file and a single character (type char). This function should open the file, count the number of times the character is found within the file, close the file, and then return the count.

If the file does not exist, this function should output an error message and then call the exit function to exit the program with an error value of 1."